+++ packages/eslint-config/src/index.ts
@@ -3,6 +3,6 @@ import config from './config';
 export * from './override-rules';
 export * from './types';
 export * from './utils';
-export type * from '@pixpilot/eslint-config';
+export type * from '@pixpilot/antfu-eslint-config';
 
 export default config;