+++ packages/eslint-config/src/types/index.ts
@@ -1,4 +1,4 @@
-import type config from '@pixpilot/eslint-config';
+import type config from '@pixpilot/antfu-eslint-config';
 
 export type ConfigFuncType = typeof config;
 export type ConfigOptions = Parameters<ConfigFuncType>[0] & { prettier?: boolean };